pc: SGX: add 'epc' machine option
Add a machine option 'epc', which accepts a size parameter
specifying the amount of virtual EPC to allocate via KVM.
Requesting a non-zero virtual EPC size will result in an
error if the necessary KVM support is not present.

KVM enforces SGX CPUID support if virtual EPC is allocated
for a VM, e.g. allocating virtual EPC but not enabled SGX
in a VCPU's CPUID will cause the KVM_SET_CPUID2 ioctl to
fail with -EINVAL.  This dependency is not enforced in Qemu
when allocating virtual EPC as the EPC is a VM resource and
must be finalized prior to creating VCPUs.

In this initial patch, virtual EPC is always allocated from
memory above 4g; support for allocating virtual EPC below
4g will be added in a future patch.  Above 4g, virtual EPC
is allocated immediately above RAM and adjusts the memory
hotplug base accordingly.

Virtual EPC is not currently compatible with a VM that has
NUMA nodes defined.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>