target-i386: SGX: block migration if VM has virtual EPC
Block migration if a VM has virtual EPC allocated.  EPC pages
are encrypted with an ephemereal key that is tied to a specific
hardware CPU, and so in-use EPC pages cannot be moved between
CPUs/platforms without losing the contents of the page.

KVM does not track a VM's individual EPC page allocations and
so doesn't know whether or not a VM is actively using its virtual
EPC, thus we cannot opportunistically support migration at this
time, e.g. migration could be allowed if we knew the VM didn't
have in-use EPC pages.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>