target-i386: SGX: block migration if virtual EPC is in use
Block migration if a VM is actively using virtual EPC.  EPC pages
are encrypted with an ephemereal key that is tied to a specific
hardware CPU, and so in-use EPC pages cannot be moved between
CPUs/platforms without losing the contents of the page.  KVM
alerts userspace to a transition in EPC activity state via a
dedicated exit type.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>